backend/requirements/_base.txt
# Django==1.6.2
https://www.djangoproject.com/download/1.7c3/tarball/
# South==0.8.4
django-angular==0.7.0
django-extensions==1.3.3
ipdb==0.8
ipython==2.0.0
six==1.6.1
wsgiref==0.1.2
psycopg2==2.5.3
django-cors-headers==0.13

